move to remote

+++ src/evaluate/kfold_EALSTM_tune2.py
@@ -147,15 +147,19 @@ for hid_ct,n_hidden in enumerate(n_hid_arr):
                                                             seq_length, n_total_feats,\
                                                             win_shift = win_shift, begin_loss_ind = begin_loss_ind,\
                                                             static_feats=True,n_static_feats = 4) 
+
+
+            np.save("ealstm_trn_data_fold"+str(k)+".npy",trn_data)
+        else:
+            trn_data = torch.from_numpy(np.load("ealstm_trn_data_fold"+str(k)+".npy"))
+        if not os.path.exists("./ealstm_tst_data_fold"+str(k)+".npy"):
             (tst_data, _) = buildLakeDataForRNN_multilakemodel_conus(test_lakenames,\
                                                         seq_length, n_total_feats,\
                                                         win_shift = win_shift, begin_loss_ind = begin_loss_ind,\
                                                         static_feats=True,n_static_feats = 4) 
-
-            np.save("ealstm_trn_data_fold"+str(k)+".npy",trn_data)
             np.save("ealstm_tst_data_fold"+str(k)+".npy",tst_data)
+
         else:
-            trn_data = torch.from_numpy(np.load("ealstm_trn_data_fold"+str(k)+".npy"))
             tst_data = torch.from_numpy(np.load("ealstm_tst_data_fold"+str(k)+".npy"))
 
         # sys.exit()
